-findPath now returns a boolean. -g,h,f costs are now float32, boosting benchmark

+++ node.go
@@ -1,11 +1,11 @@
 package astar
 
-//A node in the graph with position and cost information.
+// A node in the graph with position and cost information.
 type node struct {
 	position Coords
-	g_cost   float64
-	h_cost   float64
-	f_cost   float64
+	g_cost   float32
+	h_cost   float32
+	f_cost   float32
 	parent   *node
 	index    int // The index of the item in the heap.
 }
